--- app/search_engine/types/dict.go
@@ -1,6 +1,0 @@
-package types
-
-type DictTireTree struct {
-	Value string `json:"value"`
-	Score int64  `json:"score"`
-}
