--- app/search_engine/types/document.go
@@ -1,8 +1,0 @@
-package types
-
-// Document 文档格式
-type Document struct {
-	DocId int64  `json:"doc_id"`
-	Title string `json:"title"`
-	Body  string `json:"body"`
-}
