+++ pkg/mapreduce/map.go
@@ -1,0 +1,1 @@
+package mapreduce
