+++ pkg/mapreduce/reduce.go
@@ -1,0 +1,1 @@
+package mapreduce
