+++ types/document.go
@@ -1,5 +1,31 @@
 package types
 
+// Document 文档格式
+type Document struct {
+	DocId int64  `json:"doc_id"`
+	Title string `json:"title"`
+	Body  string `json:"body"`
+}
+
+type Data2Starrocks struct {
+	DocId int64   `json:"doc_id"`
+	Url   string  `json:"url"`
+	Title string  `json:"title"`
+	Desc  string  `json:"desc"`
+	Score float64 `json:"score"` // 质量分
+}
+
+type Task struct {
+	Columns    []string `json:"columns"`
+	BiTable    string   `json:"bi_table"`
+	SourceType int      `json:"source_type"` // 来源 1 爬虫 2 csv导入
+}
+
+type DictTireTree struct {
+	Value string `json:"value"`
+	Score int64  `json:"score"`
+}
+
 // InvertedIndexValue 倒排索引
 type InvertedIndexValue struct {
 	Token         string        `json:"token"`
